Create subcommand action that renders sub help
A new Commander::Command method `sub_command_help` has been included
that causes the subcommand help to be rendered. Now when a subcommand
is called, instead of rendering command_help.erb, subcommand_help.erb
is rendered instead. The sub_command_method by default will rexec with
the help flag as its action. Note this action can be overridden.

Runner had to be modified to detect if a Command has the sub_command_help
flag. If so, only subcommands of Command are passed into the templater.
All other commands are removed from the scope.

Help_Formatter had to be modified to allow the Command binding
to be merged with the Runner binding. This is because the full list of
subcommands are in the Runner scope where the active commands info is
in the Command scope.

Helper_Formatter::Terminal has been modified to specificly merge the Command
and Runner bindings and render the subcommand help.

Two new erb templates had to be created for the subcommand help.