Rename sub command methods for clarity
The `hidden` method has been aliased as `sub_command`. This way a
command can be marked as a subcommand in the cli making it clearer
what it's functionality is. It will also allow additional features
to be added to sub_commands but not hidden into the future.

Sub_command_help has been renamed to sub_command_group as well to mark
that it applies to the group command and not the actual sub_command.