Support default values for Command Options
When an option for a command is intialized, it now can take an optional
default value. This default value has be passed in as a hash.

When the default option is included, it will always be loaded value into
the option parser. Then it will be overrided by the value from the command
line (if included).