Add method to suppress the --trace message
There are certain errors in which the "--trace" message should not be
displayed to the user. This is to distinguish between user errors (in
which the backtrace is not useful) and all other errors (in which it
is).

This commit updates the spec and adds the method

NOTE: If the "--trace" option is included, the backtrace will still
be displayed. This method only concerns whether the option is displayed
to the user.